Got it to compile, more or less.
Also the initial min-rank is set to nrows instead of nrows + 1, to avoid any
risk of overflow. This should only have an effect in strange cases where there
are many missing values such that the initial min-rank actually survives.

Also added tests for the new utility to cap the cache size.