Skip calculation of per-combination means/variances/detected if not needed. (#8)
This provides some opportunities for optimization if we don't need certain
statistics, e.g., don't compute the variances if we don't want Cohen's d.

We add some extra options to all the score_markers_* functions to avoid
computing to the groupwise statistics if they're not of interest.

Also fleshed out the tests.