Limit the min-rank reporting in score_markers_summary for efficiency. (#9)
This allows us to accumulate the top-ranked genes from each pairwise comparison
in a priority queue, similar to score_markers_best(); and then collapse this
into a single min-rank vector that only makes sense up to the specified limit.

In doing so, we avoid the need to instantiate the 3D array for AUCs for greater
memory efficiency. We can also process simple effect sizes on a gene-by-gene
basis to eliminate the need to cache effect sizes of the reverse comparison.